Move library between diretories (#444)
